refactor: move google scholar citation endpoint
