+++ paperscraper/citations/__init__.py
@@ -1,2 +1,3 @@
-from ..scholar import get_citations_from_title
+from .citations import get_citations_by_doi, get_citations_from_title
+from .core import SelfLinkClient
 from .self_references import self_references, self_references_paper
